Update HomePage with a new HomeView
